source/checker/requirements.txt
antlr4-python3-runtime==4.13.2
asciimatics==1.15.0
blinker==1.9.0
capstone==5.0.3
click==8.1.7
dacite==1.8.1
dill==0.3.9
enum-compat==0.0.3
first==2.0.2
Flask==3.1.0
gevent==24.11.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpath-ng==1.6.1
keystone-engine==0.9.2
loguru==0.7.2
MarkupSafe==3.0.2
mpmath==1.3.0
multiprocess==0.70.17
overrides==7.7.0
pefile==2024.8.26
pillow==10.4.0
ply==3.11
prompt-toolkit==3.0.36
pyelftools==0.31
pyfiglet==1.0.2
pyperclip==1.9.0
python-fx==0.3.2
python-registry==1.3.1
PyYAML==6.0.2
qiling==1.4.6
questionary==2.0.1
termcolor==2.5.0
typing_extensions==4.12.2
unicodecsv==0.14.1
unicorn==2.1.1
urwid==2.6.15
wcwidth==0.2.13
Werkzeug==3.1.3
yamale==5.2.1
zope.event==5.0
zope.interface==7.2
